Rename project to mockapi
While this still works, the project is very lightly maintained and
almost never used by anyone. A more active project (currently called
Lusty) is interested in using this name. Strapi is a good name that
deserves an active project.
