+++ cli.js
@@ -9,20 +9,20 @@ var route = [
 ].join(os.EOL);
 
 var app = [
-    'var strapi = require(\'strapi\')'
-  , '  , express = strapi.express;'
+    'var mockapi = require(\'mockapi\')'
+  , '  , express = mockapi.express;'
   , ''
   , 'var app = express();'
-  , 'var routes = strapi.routes(\'routes\', __dirname);'
+  , 'var routes = mockapi.routes(\'routes\', __dirname);'
   , ''
   , 'app.set(\'port\', process.env.PORT || 3000);'
   , 'app.use(express.favicon());'
   , 'app.use(express.logger(\'dev\'));'
   , 'app.use(express.bodyParser());'
-  , 'app.use(strapi.middleware.forceJsonResponse());'
+  , 'app.use(mockapi.middleware.forceJsonResponse());'
   , 'app.use(app.router);'
-  , 'app.use(strapi.middleware.notFoundHandler());'
-  , 'app.use(strapi.middleware.errorHandler(app.get(\'env\')));'
+  , 'app.use(mockapi.middleware.notFoundHandler());'
+  , 'app.use(mockapi.middleware.errorHandler(app.get(\'env\')));'
   , ''
   , '// api routes'
   , 'app.get(\'/hello/:world\', routes.hello.get);'
@@ -41,7 +41,7 @@ var pkg = JSON.stringify({
     start: "node app.js"
   },
   dependencies: {
-    strapi: require('./package.json').version
+    mockapi: require('./package.json').version
   }
 });
